increase the size of the border variables from 8 to 16 bits
2007-08-22  Tommi Komulainen  <tommi.komulainen@nokia.com>

	* HACKING
	* src/sapwood-proto.h (PixbufOpenRequest)
	* src/theme-pixbuf.h (ThemePixbuf): increase the size of the border
	variables from 8 to 16 bits to allow larger borders. (NB#63819)
	NOTE: this breaks the binary protocol between client and the server


git-svn-id: https://stage.maemo.org/svn/maemo/projects/haf/trunk/sapwood@13350 50eb543b-83fa-0310-9bb7-f7cbbee4ab39

+++ src/sapwood-proto.h
@@ -39,10 +39,10 @@ typedef struct
 typedef struct
 {
   PixbufBaseRequest base;
-  guint8  border_left;
-  guint8  border_right;
-  guint8  border_top;
-  guint8  border_bottom;
+  guint16 border_left;
+  guint16 border_right;
+  guint16 border_top;
+  guint16 border_bottom;
   guchar  filename[0];          /* null terminated, absolute filename */
 } PixbufOpenRequest;
